+++ src/core/util/packed/growable_writer.rs
@@ -104,7 +104,7 @@ impl GrowableWriter {
 }
 
 impl Reader for GrowableWriter {
-    fn get(&self, index: i32) -> i64 {
+    fn get(&self, index: usize) -> i64 {
         self.current.get(index)
     }
